% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.functional.R
\name{multiplicity.functional}
\alias{multiplicity.functional}
\title{Functional Multiplicity}
\usage{
multiplicity.functional(ab, diss, ab_clust, diss_clust, sig = 1)
}
\arguments{
\item{ab}{1 x n vector of abundances before clustering.}

\item{diss}{n x n matrix with dissimilarities / distances before clustering}

\item{ab_clust}{1 x n vector of abundances after clustering.}

\item{diss_clust}{n x n matrix with dissimilarities / distances before
clustering}

\item{sig}{float value determining that two units are different}
}
\value{
FM_sigma.
}
\description{
Functional Multiplicity defined as the ratio between the functional
diversities of the unclustered set and the cluster set
}
