% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.taxonomic.R
\name{multiplicity.taxonomic}
\alias{multiplicity.taxonomic}
\title{Taxonomic Multiplicity}
\usage{
multiplicity.taxonomic(ab, clust, q = 1)
}
\arguments{
\item{ab}{List or vector element of size n with the unit's abundances.}

\item{clust}{List or vector  element of size n with the unit's corresponding
cluster}

\item{q}{q parameter for the Hill number}
}
\value{
^qM
}
\description{
Taxonomic multiplicity directly from the factor of the true diversity
unclustered abundances and the clustered abundances.
}
\examples{
ab <- rep(10,9)
clust <- c(rep(1,3), rep(2,3), rep(3,3))
multiplicity.taxonomic(ab, clust)

}
