% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.distance.R
\name{diversity.functional}
\alias{diversity.functional}
\title{Functional diversity}
\usage{
diversity.functional(ab, diss, sig = 1)
}
\arguments{
\item{ab}{1 x n vector of abundances.}

\item{diss}{n x n matrix with dissimilarities / distances}

\item{sig}{float value determining that two units are different}
}
\value{
FD_sigma.
}
\description{
Functional diversity derived from Chiu & Chao 2014, with q = 1 ans assuming
no intra specific variation. FD_sigma in the manuscript.
}
