% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_indices.R
\name{diversity.functional}
\alias{diversity.functional}
\title{Functional diversity (Chiu & Chao 2014)}
\usage{
diversity.functional(ab, diss, sig = 1)
}
\arguments{
\item{ab}{A numeric vector of element abundances.}

\item{diss}{A numeric matrix representing the dissimilarities or distances between elements.}

\item{sig}{A numeric value determining the threshold (sigma) at which two units are considered different.}
}
\value{
A numeric value representing the functional diversity, \eqn{FD_{\sigma}}.
}
\description{
Computes the functional diversity, derived from Chiu & Chao (2014),
assuming no intra-specific variation and using a default q value of 1.
}
\references{
\itemize{
\item Chiu CH, Chao A (2014) Distance-based functional diversity measures and their decomposition: A framework based on Hill numbers. PLOS ONE 9(7). \doi{10.1371/journal.pone.0100014}. \url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0100014}
}
}
