% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.distance.R
\name{diversity.functional.traditional}
\alias{diversity.functional.traditional}
\title{Functional diversity (From Chiu & Chao 2014 )}
\usage{
diversity.functional.traditional(ab, diss, q = 1)
}
\arguments{
\item{ab}{1 x n vector of abundances.}

\item{diss}{n x n matrix with dissimilarities / distances}

\item{q}{q parameter for the Hill number}
}
\value{
^qFD
}
\description{
Functional diversity derived from Chiu & Chao 2014. This corresponds to
D(Q) in their paper and ^qFD in the manusctipt
}
