% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_indices.R
\name{diversity.functional.traditional}
\alias{diversity.functional.traditional}
\title{Distance-based functional diversity (order q) (Chiu & Chao 2014)}
\usage{
diversity.functional.traditional(ab, diss, q = 1)
}
\arguments{
\item{ab}{Numeric vector of element abundances.}

\item{diss}{Numeric matrix of pairwise dissimilarities among elements.}

\item{q}{Numeric order of the Hill number (non-negative).}
}
\value{
Numeric scalar, the distance-based functional diversity \eqn{^{q}FD}{FD^q}.
}
\description{
Computes distance-based functional diversity \eqn{^{q}FD}{FD^q} following Chiu & Chao (2014).
This corresponds to D(Q) in their paper and \eqn{^{q}FD}{FD^q} in the divermeta manuscript.
For \eqn{q = 1}{q = 1}, an approximation is used internally.
}
\references{
\itemize{
\item Chiu CH, Chao A (2014) Distance-based functional diversity measures and their decomposition: A framework based on Hill numbers. PLOS ONE 9(7). \doi{10.1371/journal.pone.0100014}. \url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0100014}
}
}
\seealso{
\code{\link[=diversity.functional]{diversity.functional()}}, \code{\link[=raoQuadratic]{raoQuadratic()}}
}
