% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.distance.R
\name{multiplicity.distance}
\alias{multiplicity.distance}
\title{Distance based Multiplicity}
\usage{
multiplicity.distance(ab, diss, ab_clust, diss_clust, sig = 1)
}
\arguments{
\item{ab}{A numeric vector of element abundances before clustering.}

\item{diss}{A numeric matrix representing the dissimilarities or distances between elements before clustering.}

\item{ab_clust}{A numeric vector of element abundances after clustering.}

\item{diss_clust}{A numeric matrix representing the dissimilarities or distances between elements after clustering.}

\item{sig}{A numeric value determining the threshold (sigma) at which two units are considered different.}
}
\value{
A numeric value representing the distance multiplicity, \eqn{FM_{\sigma}}.
}
\description{
Computes the distance-based multiplicity,
defined as the ratio between the functional diversities of the unclustered set and the clustered set.
}
