% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.inventory.R
\name{multiplicity.inventory}
\alias{multiplicity.inventory}
\title{Inventory Multiplicity}
\usage{
multiplicity.inventory(ab, clust, q = 1)
}
\arguments{
\item{ab}{A numeric vector or list of element abundances.}

\item{clust}{A numeric vector or list indicating the cluster assignment for each element.}

\item{q}{A numeric parameter for the Hill number, which determines the diversity order.}
}
\value{
A numeric value representing the inventory multiplicity, \eqn{^qM}.
}
\description{
Computes the inventory multiplicity, derived from the factor of the true diversity of unclustered abundances and the clustered abundances.
}
\examples{
ab <- rep(10,9)
clust <- c(rep(1,3), rep(2,3), rep(3,3))
multiplicity.inventory(ab, clust)

}
