% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raoQ.R
\name{raoQuadratic}
\alias{raoQuadratic}
\title{Rao Quadratic}
\usage{
raoQuadratic(ab, diss)
}
\arguments{
\item{ab}{1 x n vector of abundances.}

\item{diss}{n x n matrix with dissimilarities / distances}
}
\value{
Q.
}
\description{
TODO
}
