% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raoQ.R
\name{raoQuadratic}
\alias{raoQuadratic}
\title{Rao quadratic entropy (Q) (Rao 1982)}
\usage{
raoQuadratic(ab, diss)
}
\arguments{
\item{ab}{Numeric vector of element abundances.}

\item{diss}{Numeric square matrix of pairwise dissimilarities among elements.}
}
\value{
Numeric scalar, Rao quadratic entropy \eqn{Q}{Q}.
}
\description{
Computes Rao quadratic entropy \eqn{Q}{Q}, a diversity measure that accounts for
element abundances and their pairwise dissimilarities.
}
\references{
\itemize{
\item Rao CR (1982) Diversity and dissimilarity coefficients: A unified approach. Theoretical Population Biology, 21. \doi{10.1016/0040-5809(82)90004-1}.
}
}
