% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_indices.R
\name{redundancy}
\alias{redundancy}
\title{Functional redundancy (Re) (Ricotta & Pavoine 2025)}
\usage{
redundancy(ab, diss)
}
\arguments{
\item{ab}{Numeric vector of element abundances.}

\item{diss}{Numeric square matrix or \code{dist} object of pairwise dissimilarities among elements
scaled to the range [0, 1].}
}
\value{
Numeric scalar, functional redundancy \code{Re}.
}
\description{
Computes functional redundancy \eqn{Re}{Re}, a measure of the degree to which
distinct elements are functionally similar given their abundances and
pairwise dissimilarities. This implementation follows the Simpson–Rao
family and corresponds to the \code{q = 2} case.
}
\references{
\itemize{
\item Ricotta C, Pavoine S (2025) What do functional diversity, redundancy, rarity, and originality actually measure? A theoretical guide for ecologists and conservationists. Ecological Complexity 61. \doi{10.1016/j.ecocom.2025.101116}. \url{https://www.sciencedirect.com/science/article/pii/S1476945X25000017}
\item Rao CR (1982) Diversity and dissimilarity coefficients: A unified approach. Theoretical Population Biology 21. \doi{10.1016/0040-5809(82)90004-1}.
}
}
\seealso{
\code{\link[=raoQuadratic]{raoQuadratic()}}, \code{\link[=redundancy.by_blocks]{redundancy.by_blocks()}} for computing from a compact distance table
}
